add functionalities service+specialities

+++ DataAccess/CaseServiceRepository.cs
@@ -148,14 +148,14 @@ namespace DataAccess
                             CaseService caseService = new CaseService();
                             caseService.CaseID = (int)reader["caseID"];
                             caseService.ServiceID = (int)reader["serviceID"];
-                            caseService.Hours = (decimal)reader["hours"];
-                            caseService.Kilometres = (decimal)reader["kilometres"];
-                            caseService.EstimatedHours = (decimal)reader["estimatedHours"];
+                            caseService.Hours = (float)reader["hours"];
+                            caseService.Kilometres = (float)reader["kilometres"];
+                            caseService.EstimatedHours = (float)reader["estimatedHours"];
 
                             Service service = new Service();
                             service.ID = (int)reader["ID"];
                             service.Name = (string)reader["name"];
-                            service.Price = (decimal)reader["price"];
+                            service.Price = (float)reader["price"];
                             service.IsHourly = (bool)reader["isHourly"];
 
                             servicesByCaseService.Add(caseService, service);
@@ -179,9 +179,9 @@ namespace DataAccess
         {
             caseService.CaseID = (int)reader[0];
             caseService.ServiceID = (int)reader[1];
-            caseService.Hours = (decimal)reader[2];
-            caseService.Kilometres = (decimal)reader[3];
-            caseService.EstimatedHours = (decimal)reader[3];
+            caseService.Hours = (float)reader[2];
+            caseService.Kilometres = (float)reader[3];
+            caseService.EstimatedHours = (float)reader[3];
         }
         private static List<CaseService> MapCollection(SqlCommand command)
         {
